requirements.txt
# Core dependencies
pydantic-ai>=0.0.14
python-telegram-bot>=20.0
ollama>=0.1.0
openai>=1.0.0
google-generativeai>=0.3.0

# Supporting dependencies
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
notion-client>=2.2.0
google-api-python-client>=2.100.0
google-auth>=2.23.0

# Optional dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0

